+++ opentelemetry-sdk/tests/trace/propagation/test_b3_format.py
@@ -38,7 +38,7 @@ def get_child_parent_new_carrier(old_carrier):
         "child",
         trace_api.SpanContext(
             parent_span_context.trace_id,
-            trace_api.RandomIdsGenerator().generate_span_id(),
+            RandomIdsGenerator().generate_span_id(),
             is_remote=False,
             trace_flags=parent_span_context.trace_flags,
             trace_state=parent_span_context.trace_state,